requirements.txt
numpy
matplotlib
scikit-image
scipy
opencv-python
streamlit
pillow
fastapi
uvicorn[standard]
sqlalchemy